Make 'ProcessPostcodeWorker' more resilient to new postcodes
In theory, any postcode that doesn't yet exist would either be:

1) added via the `populate_postcodes_table` rake task, i.e. given
   a placeholder result of `{}` and then updated asynchronously, or
2) added via a call to `OsPlacesApi::Client.locations_for_postcode`

However, if a developer goes into the console and calls
`ProcessPostcodeWorker.perform(postcode)` on a new postcode, it
will raise an exception `NoMethodError (undefined method destroy'
for nil:NilClass)`. We want to have the worker encounter such
scenarios gracefully. We could either have it raise an exception
along the lines of "Tried updating a postcode that hasn't been
added yet", but it is just as easy to create the postcode, which
is clearly what the developer would want in such a case.

+++ lib/os_places_api/client.rb
@@ -24,9 +24,11 @@ module OsPlacesApi
       response = get_api_response(postcode)
       record = Postcode.find_by(postcode: postcode)
       if response["results"].nil?
-        record.destroy
+        record.destroy unless record.nil?
+      elsif record.nil?
+        Postcode.create!(postcode: postcode, results: response["results"])
       else
-        record.update(results: response["results"], updated_at: Time.now)
+        record.update(results: response["results"])
       end
     end
